Refactor to avoid exporting PATH in env hook
